requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\Personal\camera_server\main.py: 3,4,5,11
fastapi == 0.58.0

# C:\Personal\camera_server\main.py: 6
# C:\Personal\camera_server\src\components\loggermanager.py: 2
loguru == 0.3.2

# C:\Personal\camera_server\src\components\models.py: 1
pydantic == 1.5.1

uvicorn == 0.11.5

python-multipart==0.0.5

RPi.GPIO==0.7.0

picamera==1.13

requests==2.24.0
